+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   listener.
 - Bug Fix: Allow canceling and re-listening to broadcast streams after a
   `merge` transform.
+- Bug Fix: Broadcast streams which are buffered using a single-subscription
+  trigger can be canceled and re-listened.
+- Bug Fix: Buffer outputs one more value if there is a pending trigger before
+  the trigger closes.
 - Bug Fix: Single-subscription streams concatted after broadcast streams are
   handled correctly.
 - Use sync `StreamControllers` for forwarding where possible.